+++ tdigest/src/Data/TDigest/Vector/NonEmpty.hs
@@ -1,0 +1,1 @@
+module Data.TDigest.Vector.NonEmpty where