+++ tdigest-Chart/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2 
+
+- compatible with `tdigest-0.2`
+
 ## 0
 
 - initial version