+++ tdigest/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2
+
+- Add `Data.TDigest.Vector` module.
+
 ## 0.1
 
 - Add `validateHistogram` and `debugPrint`